requirements_dev.txt
-i https://pypi.org/simple
alabaster==0.7.16; python_version >= '3.9'
astroid==3.2.2; python_full_version >= '3.8.0'
babel==2.15.0; python_version >= '3.8'
black==24.4.2; python_version >= '3.8'
certifi==2024.6.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
codacy-coverage==1.3.11
codecov==2.1.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
colorama==0.4.6; sys_platform == 'win32'
coverage[toml]==7.5.3; python_version >= '3.8'
dill==0.3.8; python_version >= '3.11'
doc8==1.1.1; python_version >= '3.8'
docutils==0.20.1; python_version >= '3.7'
flake8==7.1.0; python_full_version >= '3.8.1'
idna==3.7; python_version >= '3.5'
imagesize==1.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==7.2.0; python_version >= '3.8'
iniconfig==2.0.0; python_version >= '3.7'
isort==5.13.2; python_full_version >= '3.8.0'
jaraco.classes==3.4.0; python_version >= '3.8'
jaraco.context==5.3.0; python_version >= '3.8'
jaraco.functools==4.0.1; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
keyring==25.2.1; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mccabe==0.7.0; python_version >= '3.6'
mdurl==0.1.2; python_version >= '3.7'
more-itertools==10.3.0; python_version >= '3.8'
mypy-extensions==1.0.0; python_version >= '3.5'
nh3==0.2.17
packaging==24.1; python_version >= '3.8'
pathspec==0.12.1; python_version >= '3.8'
pbr==6.0.0; python_version >= '2.6'
pkginfo==1.11.1; python_version >= '3.8'
platformdirs==4.2.2; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
pycodestyle==2.12.0; python_version >= '3.8'
pyflakes==3.2.0; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pylint==3.2.3; python_full_version >= '3.8.0'
pytest==8.2.2; python_version >= '3.8'
pytest-cov==5.0.0; python_version >= '3.8'
pywin32-ctypes==0.2.2; sys_platform == 'win32'
pyyaml==6.0.2rc1; python_version >= '3.6'
readme-renderer==43.0; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
restructuredtext-lint==1.4.0
rfc3986==2.0.0; python_version >= '3.7'
rich==13.7.1; python_full_version >= '3.7.0'
snowballstemmer==2.2.0
sphinx==7.3.7; python_version >= '3.9'
sphinxcontrib-applehelp==1.0.8; python_version >= '3.9'
sphinxcontrib-devhelp==1.0.6; python_version >= '3.9'
sphinxcontrib-htmlhelp==2.0.5; python_version >= '3.9'
sphinxcontrib-jsmath==1.0.1; python_version >= '3.5'
sphinxcontrib-qthelp==1.0.7; python_version >= '3.9'
sphinxcontrib-serializinghtml==1.1.10; python_version >= '3.9'
stevedore==5.2.0; python_version >= '3.8'
tomlkit==0.12.5; python_version >= '3.7'
twine==5.1.0; python_version >= '3.8'
urllib3==2.2.2; python_version >= '3.8'
zipp==3.19.2; python_version >= '3.8'
cffi==1.17.0rc1; python_version >= '3.8'
clr-loader==0.2.6; python_version >= '3.7'
numpy==1.26.4; python_version >= '3.9'
pandas==2.2.2; python_version >= '3.9'
-e . ; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pythonnet==3.0.3; python_version < '3.13' and python_version >= '3.7'
pytz==2024.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tzdata==2024.1; python_version >= '2'
wrapt==1.16.0; python_version >= '3.6'
